Migrate lua helpers to datafile parser
