+++ src/modules/lua/helpers.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include "modules/engine/render.h"
 #include "modules/site/connector.h"
 #include "modules/stats/stats.h"
 #include <flecs.h>
@@ -14,34 +13,26 @@ struct RocketName {
 struct RocketPrefabType {
   std::string value;
 };
-struct TextureName {
-  std::string value;
-};
-struct TextureFilename {
-  std::string value;
-};
-struct TextureModName {
-  std::string value;
-};
-struct SpriteClipRect {
-  int x;
-  int y;
-  uint32_t width;
-  uint32_t height;
-};
 
 void load_helpers_namespace(lua_State *L);
+
+/**
+ * @brief Create (or reuse, if it already exists) a named prefab under `parent`.
+ *
+ * Scoping the creation makes the name resolve relative to `parent`, so a repeat
+ * call with the same name returns the existing prefab instead of creating a
+ * conflicting root entity. This keeps prefab registration idempotent, which the
+ * developer "Reload Prefabs" path relies on.
+ */
+flecs::entity create_prefab_under(const flecs::world &world,
+                                  flecs::entity parent, flecs::entity base,
+                                  const std::string &name);
 
 flecs::entity create_site(const flecs::world &world, const std::string &name,
                           uint8_t width, uint8_t height,
                           bool make_current = false);
-flecs::entity create_building_prefab(const flecs::world &world,
-                                     const std::string &name);
 flecs::entity create_rocket_prefab(const flecs::world &world,
                                    const std::string &name);
-flecs::entity add_facility_to_building(const flecs::world &world,
-                                       flecs::entity building,
-                                       const std::string &name);
 flecs::entity create_rocket(const flecs::world &world, const RocketName &name,
                             const RocketPrefabType &prefab,
                             flecs::entity parent = flecs::entity());
@@ -52,16 +43,10 @@ flecs::entity add_target_orbit_to_rocket(const flecs::world &world,
                                          flecs::entity rocket,
                                          const std::string &orbit_name,
                                          uint32_t max_mass);
-flecs::entity create_texture(flecs::world world, const TextureName &name,
-                             TextureFilename filename,
-                             const TextureModName &mod_name);
 flecs::entity create_effect(const flecs::world &world, const std::string &name,
                             flecs::entity source);
 flecs::entity add_modifier(const flecs::world &world, flecs::entity effect,
                            const Modifier &mod);
-Sprite clip_sprite_from_texture(const flecs::world &world,
-                                const std::string &texture,
-                                SpriteClipRect rect);
 flecs::entity create_contract(flecs::world &world, const std::string &name,
                               const std::string &client,
                               const std::string &description,
