+++ src/modules/lua/mod_content.h
@@ -13,6 +13,7 @@
  */
 #pragma once
 
+#include "modules/engine/render.h" // Sprite
 #include "modules/lua/building_data.h"
 #include "modules/lua/texture_data.h"
 #include <flecs.h>
@@ -37,6 +38,49 @@ struct ReloadPrefabsRequest {};
  * unrecognised type, leaving the facility untouched.
  */
 bool addFacilityComponent(flecs::entity facility, const std::string &type);
+
+// Strong typedefs so the three string arguments to create_texture cannot be
+// transposed at a call site (enforced by clang-tidy's swappable-parameters
+// check).
+struct TextureName {
+  std::string value;
+};
+struct TextureFilename {
+  std::string value;
+};
+struct TextureModName {
+  std::string value;
+};
+
+/**
+ * @brief Load `mods/<mod_name>/<filename>` as an SDL texture under Textures::.
+ *
+ * Rejects filenames containing "..". Idempotent: a repeat call reuses the
+ * existing `Textures::<name>` entity and overwrites its Texture, so reloading
+ * mod content does not create conflicting root entities.
+ */
+flecs::entity create_texture(flecs::world world, const TextureName &name,
+                             const TextureFilename &filename,
+                             const TextureModName &mod_name);
+
+/// @brief Create (or reuse) an empty building prefab under Prefabs::Buildings.
+flecs::entity create_building_prefab(const flecs::world &world,
+                                     const std::string &name);
+
+/// @brief Create (or reuse) a facility prefab child of `building`.
+flecs::entity add_facility_to_building(const flecs::world &world,
+                                       flecs::entity building,
+                                       const std::string &name);
+
+/**
+ * @brief Build a Sprite that clips `rect` out of the named
+ * `Textures::<texture>`.
+ *
+ * Returns an empty Sprite (and logs) if the texture does not exist.
+ */
+Sprite clip_sprite_from_texture(const flecs::world &world,
+                                const std::string &texture,
+                                SpriteClipRect rect);
 
 /// @brief Load each parsed texture for `mod_name` into the world.
 void applyTextureData(flecs::world &world, const std::string &mod_name,
