+++ test/CMakeLists.txt
@@ -23,16 +23,12 @@ add_executable(unit_test
   lua/entity/entity_lookup.test.cpp
   lua/entity/entity_is_a.test.cpp
   lua/helpers/create_site.test.cpp
-  lua/helpers/create_building_prefab.test.cpp
   lua/helpers/create_rocket_prefab.test.cpp
-  lua/helpers/add_facility_to_building.test.cpp
   lua/helpers/create_rocket.test.cpp
   lua/helpers/create_building.test.cpp
   lua/helpers/add_target_orbit_to_rocket.test.cpp
-  lua/helpers/create_texture.test.cpp
   lua/helpers/create_effect.test.cpp
   lua/helpers/add_modifier.test.cpp
-  lua/helpers/clip_sprite_from_texture.test.cpp
   lua/helpers/create_contract.test.cpp
   lua/helpers/create_contract_payload.test.cpp
   lua/helpers/get_all_contracts.test.cpp
@@ -48,6 +44,10 @@ add_executable(unit_test
   lua/data/building_data.test.cpp
   lua/data/texture_data.test.cpp
   lua/mod_content/apply_building_data.test.cpp
+  lua/mod_content/create_texture.test.cpp
+  lua/mod_content/create_building_prefab.test.cpp
+  lua/mod_content/add_facility_to_building.test.cpp
+  lua/mod_content/clip_sprite_from_texture.test.cpp
   rocket/systems.test.cpp
   rocket/actions/cancel_launch_action.test.cpp
   rocket/actions/launch_initiate_rollout_action.test.cpp
