+++ test/lua/helpers/create_building.test.cpp
@@ -1,4 +1,5 @@
 #include "modules/lua/helpers.h"
+#include "modules/lua/mod_content.h"
 #include "setup_helpers.h"
 #include <catch2/catch_test_macros.hpp>
 #include <flecs.h>
