+++ test/lua/mod_content/create_building_prefab.test.cpp
@@ -1,5 +1,5 @@
-#include "modules/lua/helpers.h"
-#include "setup_helpers.h"
+#include "../helpers/setup_helpers.h"
+#include "modules/lua/mod_content.h"
 #include <catch2/catch_test_macros.hpp>
 #include <flecs.h>
